requirements.txt
psycopg2==2.9.10
tkcalendar==1.6.1
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3